Renamed BGL solver files
