+++ descartes_light/solvers/CMakeLists.txt
@@ -25,7 +25,7 @@ endif()
 
 add_library(
   ${PROJECT_NAME}_solvers
-  src/bgl/bgl_ladder_graph_solver.cpp
+  src/bgl/bgl_solver.cpp
   src/ladder_graph/ladder_graph_solver.cpp
   src/ladder_graph/ladder_graph.cpp
   src/ladder_graph/ladder_graph_dag_search.cpp)
