+++ descartes_light/solvers/include/descartes_light/solvers/bgl/impl/bgl_solver.hpp
@@ -25,7 +25,7 @@ DESCARTES_IGNORE_WARNINGS_PUSH
 #include <omp.h>
 DESCARTES_IGNORE_WARNINGS_POP
 
-#include <descartes_light/solvers/bgl/bgl_ladder_graph_solver.h>
+#include <descartes_light/solvers/bgl/bgl_solver.h>
 #include <descartes_light/solvers/bgl/utils.h>
 #include <descartes_light/types.h>
