+++ descartes_light/test/include/descartes_light/test/impl/solver_factory.hpp
@@ -2,7 +2,7 @@
 
 #include <descartes_light/test/solver_factory.h>
 #include <descartes_light/solvers/ladder_graph/ladder_graph_solver.h>
-#include <descartes_light/solvers/bgl/bgl_ladder_graph_solver.h>
+#include <descartes_light/solvers/bgl/bgl_solver.h>
 
 namespace descartes_light
 {
